Remove future from the dependency list
As of c2e701c, Python2 is no longer supported, and the "future"
package, which formerly provided a compatibility layer, has been
dropped.

However, "future" is still listed as a dependency in setup.py, despite
not being used, (at least as far as I can tell).  This commit drops this
unnecessary dependency.